app/requirements.txt
Flask==1.1.1
tensorflow_serving_api==1.4.0
Flask_API==1.1
grpcio==1.26.0
tensorflow==1.14.0
scikit_image==0.14.5
six==1.11.0
numpy==1.16.6
Pillow==6.2.2
PyYAML==5.3
scikit_learn==0.20.4
flask_login
